Tidy up evaluation package
