+++ src/main/java/org/chainoptim/features/evaluation/production/connection/service/FactoryStageConnectionService.java
@@ -1,7 +1,7 @@
-package org.chainoptim.features.evaluation.production.services;
+package org.chainoptim.features.evaluation.production.connection.service;
 
-import org.chainoptim.features.evaluation.production.model.FactoryStageConnection;
-import org.chainoptim.features.evaluation.production.repository.FactoryStageConnectionRepository;
+import org.chainoptim.features.evaluation.production.connection.model.FactoryStageConnection;
+import org.chainoptim.features.evaluation.production.connection.repository.FactoryStageConnectionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
