+++ src/main/java/org/chainoptim/features/evaluation/production/evaluation/service/FactoryGraphEvaluationService.java
@@ -1,0 +1,8 @@
+package org.chainoptim.features.evaluation.production.evaluation.service;
+
+import org.chainoptim.features.evaluation.production.resourceallocation.model.AllocationPlan;
+
+public interface FactoryGraphEvaluationService {
+    AllocationPlan evaluateFactory(Integer factoryId, Float duration);
+    void analyzeFactoryGraphWithPipelines(Integer factoryId, Float duration);
+}
