+++ src/main/java/org/chainoptim/features/evaluation/production/graph/controller/GraphController.java
@@ -1,27 +1,19 @@
 package org.chainoptim.features.evaluation.production.graph.controller;
 
-import org.chainoptim.features.evaluation.production.graph.model.FactoryGraph;
-import org.chainoptim.features.evaluation.production.model.FactoryStageConnection;
-import org.chainoptim.features.evaluation.production.services.GraphService;
-import org.chainoptim.features.factory.model.Factory;
+import org.chainoptim.features.evaluation.production.graph.service.FactoryGraphService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/graphs")
 public class GraphController {
 
-    private final GraphService graphService;
+    private final FactoryGraphService factoryGraphService;
 
     @Autowired
-    public GraphController(GraphService graphService) {
-        this.graphService = graphService;
+    public GraphController(FactoryGraphService factoryGraphService) {
+        this.factoryGraphService = factoryGraphService;
     }
 
 //    @GetMapping("/{factoryId}")
