+++ src/main/java/org/chainoptim/features/evaluation/production/graph/model/FactoryGraph.java
@@ -1,9 +1,8 @@
 package org.chainoptim.features.evaluation.production.graph.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.chainoptim.features.evaluation.production.model.FactoryStageConnection;
+import org.chainoptim.features.evaluation.production.connection.model.FactoryStageConnection;
 import org.chainoptim.features.factory.model.Factory;
 import org.chainoptim.features.factory.model.FactoryStage;
 import org.chainoptim.features.productpipeline.model.StageInput;
