--- src/main/java/org/chainoptim/features/evaluation/production/services/GraphUtilsService.java
@@ -1,15 +1,0 @@
-package org.chainoptim.features.evaluation.production.services;
-
-import org.chainoptim.features.evaluation.production.graph.model.Edge;
-import org.chainoptim.features.evaluation.production.graph.model.FactoryGraph;
-import org.chainoptim.features.evaluation.production.graph.model.Node;
-import org.springframework.stereotype.Service;
-
-import java.util.*;
-
-@Service
-public class GraphUtilsService {
-
-
-
-}
