+++ src/main/java/org/chainoptim/features/factory/repository/FactoryRepository.java
@@ -1,7 +1,7 @@
 package org.chainoptim.features.factory.repository;
 
 import org.chainoptim.features.factory.model.Factory;
-import org.chainoptim.features.evaluation.production.model.FactoryStageConnection;
+import org.chainoptim.features.evaluation.production.connection.model.FactoryStageConnection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
