+++ src/main/java/org/chainoptim/features/factory/service/FactoryServiceImpl.java
@@ -6,7 +6,7 @@ import org.chainoptim.features.factory.dto.FactoriesSearchDTO;
 import org.chainoptim.features.factory.dto.FactoryDTOMapper;
 import org.chainoptim.features.factory.dto.UpdateFactoryDTO;
 import org.chainoptim.features.factory.model.Factory;
-import org.chainoptim.features.evaluation.production.model.FactoryStageConnection;
+import org.chainoptim.features.evaluation.production.connection.model.FactoryStageConnection;
 import org.chainoptim.features.factory.repository.FactoryRepository;
 import org.chainoptim.features.productpipeline.model.Stage;
 import org.chainoptim.shared.sanitization.EntitySanitizerService;
